feat: REST API 작업 추가중
- 컨트롤러 뼈대
- 화면 div active
- [ ] 화면 REST API fetch
- [ ] 화면 innerHtml
