+++ src/main/java/com/outsider/masterofprediction/controller/api/MyPageRestAPIController.java
@@ -1,0 +1,27 @@
+package com.outsider.masterofprediction.controller.api;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.ModelAndView;
+@RestController
+@RequestMapping("/api")
+public class MyPageRestAPIController {
+
+    @GetMapping("/purchase-history/{page}")
+    public ModelAndView getPurchaseHistory(@PathVariable int page, ModelAndView mv) {
+
+        return mv;
+    }
+    @GetMapping("/comments/{page}")
+    public ModelAndView comment(@PathVariable int page, ModelAndView mv) {
+
+        return mv;
+    }
+    @GetMapping("inquirys/{page}")
+    public ModelAndView getInquirys(@PathVariable int page, ModelAndView mv) {
+
+        return mv;
+    }
+}
