+++ src/main/resources/static/scss/my-page/mypage.scss
@@ -279,7 +279,9 @@ body {
     flex: 1 1 auto;
   }
   }
-
+  .active{
+    display: block !important;
+  }
   .tabpanel {
     margin-top: 0.25rem;
     padding-left: 2rem;
@@ -287,6 +289,7 @@ body {
     border-radius: 0.375rem;
     box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
     outline: none;
+    display: none;
 
     &:focus-visible {
       box-shadow: 0 0 0 2px $ring-color;
